feat: dynamic BlockchainModule
- Refactor BlockchainModule into a DynamicModule capable of providing different providers based on a readOnly/writeable deployment model
    - Required making it a Global module to resolve NestJS dependency injection issues
    - backfit to services already migrated to the common blockchain module
- Migrate content-watcher and graph services to common blockchain module
