+++ apps/account-worker/src/transaction_notifier/notifier.module.ts
@@ -1,9 +1,8 @@
 import { Module } from '@nestjs/common';
-import { BlockchainModule } from '#blockchain/blockchain.module';
 import { TxnNotifierService } from './notifier.service';
 
 @Module({
-  imports: [BlockchainModule],
+  imports: [],
   providers: [TxnNotifierService],
   exports: [],
 })
