+++ apps/account-worker/src/transaction_publisher/publisher.module.ts
@@ -1,11 +1,10 @@
 import '@frequency-chain/api-augment';
 
 import { Module } from '@nestjs/common';
-import { BlockchainModule } from '#blockchain/blockchain.module';
 import { TransactionPublisherService } from './publisher.service';
 
 @Module({
-  imports: [BlockchainModule],
+  imports: [],
   providers: [TransactionPublisherService],
   exports: [TransactionPublisherService],
 })
