+++ apps/content-publishing-worker/src/batch_announcer/batch.announcer.module.ts
@@ -1,11 +1,10 @@
 import { Module } from '@nestjs/common';
 import { BatchAnnouncementService } from './batch.announcer.service';
 import { BatchAnnouncer } from './batch.announcer';
-import { BlockchainModule } from '#blockchain/blockchain.module';
 import { IpfsService } from '#content-publishing-lib/utils/ipfs.client';
 
 @Module({
-  imports: [BlockchainModule],
+  imports: [],
   controllers: [],
   providers: [BatchAnnouncementService, BatchAnnouncer, IpfsService],
   exports: [BatchAnnouncementService, BatchAnnouncer, IpfsService],
