+++ apps/content-publishing-worker/src/publisher/publisher.module.ts
@@ -4,7 +4,7 @@ import { IPFSPublisher } from './ipfs.publisher';
 import { BlockchainModule } from '#blockchain/blockchain.module';
 
 @Module({
-  imports: [BlockchainModule],
+  imports: [],
   controllers: [],
   providers: [PublishingService, IPFSPublisher],
   exports: [PublishingService, IPFSPublisher],
