+++ apps/graph-api/src/api.module.ts
@@ -6,23 +6,24 @@ import { HealthController } from './controllers/health.controller';
 import { ApiService } from './api.service';
 import { GraphQueues as QueueConstants } from '#types/constants/queue.constants';
 import { WebhooksControllerV1 } from './controllers/v1/webhooks-v1.controller';
-import { BlockchainModule } from '#graph-lib/blockchain';
+import { BlockchainModule } from '#blockchain/blockchain.module';
 import { GraphStateManager } from '#graph-lib/services/graph-state-manager';
 import { CacheModule } from '#cache/cache.module';
 import cacheConfig, { ICacheConfig } from '#cache/cache.config';
 import { ConfigModule } from '@nestjs/config';
 import apiConfig from './api.config';
-import { allowReadOnly } from '#graph-lib/blockchain/blockchain.config';
-import queueConfig, { QueueModule } from '#queue';
+import { noProviderBlockchainConfig } from '#blockchain/blockchain.config';
+import queueConfig from '#queue';
 import scannerConfig from '#graph-worker/graph_notifier/scanner.config';
 import { AsyncDebouncerService } from '#graph-lib/services/async_debouncer';
 import graphCommonConfig from '#config/graph-common.config';
+import { QueueModule } from '#queue/queue.module';
 
 @Module({
   imports: [
     ConfigModule.forRoot({
       isGlobal: true,
-      load: [apiConfig, graphCommonConfig, allowReadOnly, cacheConfig, queueConfig, scannerConfig],
+      load: [apiConfig, graphCommonConfig, noProviderBlockchainConfig, cacheConfig, queueConfig, scannerConfig],
     }),
     EventEmitterModule.forRoot({
       // Use this instance throughout the application
@@ -42,7 +43,7 @@ import graphCommonConfig from '#config/graph-common.config';
       // disable throwing uncaughtException if an error event is emitted and it has no listeners
       ignoreErrors: false,
     }),
-    BlockchainModule,
+    BlockchainModule.forRootAsync({ readOnly: true }),
     CacheModule.forRootAsync({
       useFactory: (cacheConf: ICacheConfig) => [
         {
