+++ apps/graph-worker/src/graph_notifier/graph.monitor.processor.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { GraphMonitorService } from './graph.monitor.service';
-import { BlockchainModule } from '#graph-lib/blockchain';
+import { BlockchainModule } from '#blockchain/blockchain.module';
 import { GraphStateManager } from '#graph-lib/services/graph-state-manager';
 
 @Module({
