+++ apps/graph-worker/src/graph_notifier/graph.monitor.service.ts
@@ -5,14 +5,14 @@ import { Queue } from 'bullmq';
 import Redis from 'ioredis';
 import { MILLISECONDS_PER_SECOND } from 'time-constants';
 import { RegistryError } from '@polkadot/types/types';
-import { BlockchainService } from '#graph-lib/blockchain/blockchain.service';
+import { BlockchainService } from '#blockchain/blockchain.service';
 import { BlockchainScannerService } from '#graph-lib/utils/blockchain-scanner.service';
 import { SchedulerRegistry } from '@nestjs/schedule';
 import { SignedBlock } from '@polkadot/types/interfaces';
 import { FrameSystemEventRecord, PalletSchemasSchemaVersionId } from '@polkadot/types/lookup';
 import { HexString } from '@polkadot/util/types';
 import { IGraphTxStatus } from '#types/interfaces';
-import { CapacityCheckerService } from '#graph-lib/blockchain/capacity-checker.service';
+import { CapacityCheckerService } from '#blockchain/capacity-checker.service';
 import * as RedisConstants from '#graph-lib/utils/redis';
 import { GraphQueues as QueueConstants } from '#types/constants/queue.constants';
 import * as GraphServiceWebhook from '#graph-lib/types/webhook-types';
