+++ apps/graph-worker/src/graph_publisher/graph.publisher.processor.module.ts
@@ -1,11 +1,9 @@
 import { Module } from '@nestjs/common';
 import { GraphUpdatePublisherService } from './graph.publisher.processor.service';
-import { NonceService } from '#graph-lib/services/nonce.service';
-import { BlockchainModule } from '#graph-lib/blockchain';
 
 @Module({
-  imports: [BlockchainModule],
-  providers: [GraphUpdatePublisherService, NonceService],
-  exports: [GraphUpdatePublisherService, NonceService],
+  imports: [],
+  providers: [GraphUpdatePublisherService],
+  exports: [GraphUpdatePublisherService],
 })
 export class GraphUpdatePublisherModule {}
