+++ apps/graph-worker/src/request_processor/request.processor.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { RequestProcessorService } from './request.processor.service';
-import { BlockchainModule } from '#graph-lib/blockchain';
 import { GraphStateManager } from '#graph-lib/services/graph-state-manager';
+import { BlockchainModule } from '#blockchain/blockchain.module';
 
 @Module({
   imports: [BlockchainModule],
