+++ env-files/graph.template.env
@@ -31,8 +31,6 @@ CAPACITY_LIMIT='{"serviceLimit":{"type":"percentage","value":"80"}}'
 # Mainnet is also used for dsnp/instant-seal-node-with-deployed-schemas:latest
 GRAPH_ENVIRONMENT_TYPE=Mainnet
 
-PROVIDER_ACCOUNT_SEED_PHRASE="//Ferdie"
-
 # How many seconds to delay between successive scans of the chain
 # (after end of chain is reached)
 BLOCKCHAIN_SCAN_INTERVAL_SECONDS=6
