+++ libs/blockchain/src/blockchain.module.ts
@@ -1,13 +1,50 @@
-import '@frequency-chain/api-augment';
-
-import { Module } from '@nestjs/common';
-import { BlockchainService } from './blockchain.service';
+import { DynamicModule, Module, Provider } from '@nestjs/common';
 import { CapacityCheckerService } from './capacity-checker.service';
+import { BlockchainRpcQueryService } from './blockchain-rpc-query.service';
 
-@Module({
-  imports: [],
-  controllers: [],
-  providers: [BlockchainService, CapacityCheckerService],
-  exports: [BlockchainService, CapacityCheckerService],
-})
-export class BlockchainModule {}
+export interface IBlockchainModuleOptions {
+  readOnly: boolean;
+}
+
+@Module({})
+export class BlockchainModule {
+  static isConfigured = false;
+
+  static services: Provider[];
+
+  static async forRootAsync({ readOnly }: IBlockchainModuleOptions = { readOnly: false }): Promise<DynamicModule> {
+    const { BlockchainService } = readOnly ? { BlockchainService: undefined } : await import('./blockchain.service');
+    const BlockchainRpcQueryProvider: Provider = readOnly
+      ? {
+          provide: BlockchainRpcQueryService,
+          useClass: BlockchainRpcQueryService,
+        }
+      : { provide: BlockchainRpcQueryService, useExisting: BlockchainService };
+
+    BlockchainModule.services = [
+      CapacityCheckerService,
+      BlockchainRpcQueryProvider,
+      ...(readOnly ? [] : [BlockchainService]),
+    ];
+    BlockchainModule.isConfigured = true;
+
+    return {
+      global: true,
+      module: BlockchainModule,
+      providers: BlockchainModule.services,
+      exports: BlockchainModule.services,
+    };
+  }
+
+  // static forFeature(): DynamicModule {
+  //   // if (!BlockchainModule.isConfigured) {
+  //   //   throw new Error('Must call forRootAsync() before forFeature()');
+  //   // }
+
+  //   return {
+  //     module: BlockchainModule,
+  //     providers: BlockchainModule.services,
+  //     exports: BlockchainModule.services,
+  //   };
+  // }
+}
