+++ libs/content-watcher-lib/src/crawler/crawler.service.ts
@@ -10,6 +10,7 @@ import { BaseConsumer } from '../utils/base-consumer';
 import { ContentSearchRequestDto } from '#types/dtos/content-watcher/content-search-request.dto';
 import { ChainEventProcessorService } from '../utils/chain-event-processor.service';
 import { BlockchainService } from '#blockchain/blockchain.service';
+import { BlockchainRpcQueryService } from '#blockchain/blockchain-rpc-query.service';
 
 const CRAWLER_BLOCK_CHUNK_SIZE = 500;
 
@@ -22,7 +23,7 @@ export class CrawlerService extends BaseConsumer {
     @InjectRedis() private readonly cache: Redis,
     @InjectQueue(QueueConstants.IPFS_QUEUE) private readonly ipfsQueue: Queue,
     private readonly chainEventService: ChainEventProcessorService,
-    private readonly blockchainService: BlockchainService,
+    private readonly blockchainService: BlockchainRpcQueryService,
   ) {
     super();
   }
