+++ libs/content-watcher-lib/src/scanner/scanner.module.ts
@@ -6,7 +6,7 @@ import { BlockchainModule } from '#blockchain/blockchain.module';
 import { ChainEventProcessorModule } from '#content-watcher-lib/utils/chain-event-process.module';
 
 @Module({
-  imports: [BlockchainModule, ScheduleModule, ChainEventProcessorModule],
+  imports: [ScheduleModule, ChainEventProcessorModule],
   controllers: [],
   providers: [ScannerService],
   exports: [ScannerService],
