+++ libs/content-watcher-lib/src/utils/chain-event-process.module.ts
@@ -3,7 +3,7 @@ import { Module } from '@nestjs/common';
 import { ChainEventProcessorService } from './chain-event-processor.service';
 
 @Module({
-  imports: [BlockchainModule],
+  imports: [],
   providers: [ChainEventProcessorService],
   exports: [ChainEventProcessorService],
 })
