--- libs/graph-lib/src/blockchain/index.ts
@@ -1,5 +1,0 @@
-export * from './blockchain.module';
-export * from './blockchain.service';
-export * from './create-keys';
-export * from './event-error';
-export * from './extrinsic';
