+++ libs/queue/src/index.ts
@@ -2,4 +2,3 @@
 export { default } from './queue.config';
 export * from './queue.config';
 export * from './queue.interfaces';
-export * from './queue.module';
